Re-adding logic to rely on place detail geometry>location>lat/lng values before falling back to formatted address geocoding. Also improving error reporting and fixing warnings due to type conversion issues and logical errors.

+++ SPGooglePlacesAutocomplete/SPGooglePlacesAutocompleteQuery.m
@@ -53,7 +53,7 @@
                                                              [input stringByAddingPercentEscapesUsingEncoding:NSUTF8StringEncoding],
                                                              SPBooleanStringForBool(sensor), key];
     if (offset != NSNotFound) {
-        [url appendFormat:@"&offset=%u", offset];
+        [url appendFormat:@"&offset=%lu", (unsigned long)offset];
     }
     if (location.latitude != -1) {
         [url appendFormat:@"&location=%f,%f", location.latitude, location.longitude];
